refactoring notebook tests
